backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
httpx==0.26.0
freshrss-api==2.0.2
celery==5.5.3
pillow==11.2.1
redis==6.2.0
pydantic==2.11.5
pydantic-settings==2.2.1
starlette>=0.36.3,<0.37.0
psycopg2-binary==2.9.9
python-multipart==0.0.20
loguru==0.7.2
beartype==0.16.4
click>=8.0.0
kombu==5.5.4
vine==5.1.0
billiard==4.2.1
amqp==5.3.1
anyio==4.9.0
h11==0.16.0
sniffio==1.3.1
websockets==15.0.0
watchfiles==1.0.5
uvloop==0.21.0
casdoor==1.29.0
python-jose==3.3.0
beautifulsoup4>=4.13.4
trafilatura==1.6.1
aiohttp>=3.10.11
asyncpg==0.28.0
alembic==1.12.0
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
cryptography~=44.0.0
flower==2.0.1
